collection-service/tests/test-requirements.txt
# Test requirements
pytest
coverage
moto
python-jose
requests_mock

# Function requirements
python-jose
boto3
aws_xray_sdk
